moving parser into subpackage
